#165 extract out game creation logic to newly simplified model classes, move those into service directories
