+++ src/backend/services/chess/ChessGame.py
@@ -1,11 +1,27 @@
-
 import json
+import random
 from datetime import datetime
 
 from psycopg.types.json import Json
 
+from src.backend.pgdb import getPgdb
 from src.backend.utils import generateId
 
+pgdb = getPgdb()
+
+def createChessGame(player_name, opponent_name):
+
+    if random.choice(["Heads", "Tails"]) == "Heads":
+        white_player = player_name
+        black_player = opponent_name
+
+    else:
+        white_player = opponent_name
+        black_player = player_name
+
+    game = newChessGame(white_player, black_player)
+
+    pgdb.createChessGame(game)
 
 def newChessGame(white_player, black_player):
     boardstate = json.loads(open('resources/initialChessLayout.json', 'r').read())
